pyproject.toml
[tool.poetry]
name = "project-url-fetch"
version = "0.1.0"
description = ""
authors = ["Hubertrum <leigh.sykes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.4"
python-dateutil = "^2.9.0.post0"
jellyfish = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
